removing the dymd dependency entirely
