+++ x/gamm/types/params.go
@@ -5,8 +5,6 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-
-	apptesting "github.com/osmosis-labs/osmosis/v15/testutils/apptesting"
 )
 
 // Parameter store keys.
@@ -34,7 +32,7 @@ func NewParams(poolCreationFee sdk.Coins) Params {
 // default gamm module parameters.
 func DefaultParams() Params {
 	return Params{
-		PoolCreationFee:      sdk.Coins{sdk.NewInt64Coin(apptesting.BaseDenom, 1000_000_000)},
+		PoolCreationFee:      sdk.Coins{sdk.NewInt64Coin(sdk.DefaultBondDenom, 1000_000_000)},
 		EnableGlobalPoolFees: false,
 		GlobalFees:           GlobalFees{sdk.MustNewDecFromStr("0.02"), sdk.ZeroDec()},
 		TakerFee:             sdk.MustNewDecFromStr("0.01"),
